+++ Tests/Runtime/ReviewPopupTests.cs
@@ -1,0 +1,28 @@
+using System.Collections;
+using NUnit.Framework;
+using UnityEngine.TestTools;
+
+namespace Agava.YandexGames.Tests
+{
+    public class ReviewPopupTests
+    {
+        [UnitySetUp]
+        public IEnumerator InitializeSdk()
+        {
+            if (!YandexGamesSdk.IsInitialized)
+                yield return YandexGamesSdk.Initialize(SdkTests.TrackSuccessCallback);
+        }
+
+        [Test]
+        public void CanReviewShouldNotThrow()
+        {
+            Assert.DoesNotThrow(() => ReviewPopup.CanOpen((result, reason) => { }));
+        }
+
+        [Test]
+        public void ReviewShouldNotThrow()
+        {
+            Assert.DoesNotThrow(() => ReviewPopup.Open());
+        }
+    }
+}